Move component SevereCaseCard from 'components/cards' to 'components/index/CardsMonitoring'
