+++ components/index/CardsMonitoring.vue
@@ -49,7 +49,7 @@ import TokyoRulesApplicationNumberCard from '@/components/index/CardsMonitoring/
 // モニタリング項目(6)入院患者数
 import HospitalizedNumberCard from '@/components/index/CardsMonitoring/HospitalizedNumberCard.vue'
 // モニタリング項目(7)重症患者数
-import SevereCaseCard from '@/components/cards/SevereCaseCard.vue'
+import SevereCaseCard from '@/components/index/CardsMonitoring/SevereCaseCard.vue'
 /* eslint-enable simple-import-sort/sort */
 
 export default Vue.extend({
