+++ components/index/CardsMonitoring/SevereCaseCard.vue
